Create collatz.c
The Collatz conjecture is applied to positive integers and speculates that it is always possible to get "back to 1" if you follow these steps: 
if the previous term is even, the next term is one half the previous term. If the previous term is odd, the next term is 3 times the previous term plus 1. The conjecture is that no matter what value of n, the sequence will always reach 1.